Fix the data type problem for nkids in ageing module

+++ minos/modules/ageing.py
@@ -61,7 +61,7 @@ class Ageing(Base):
         logging.info(f"Aged population to year {event.time.year}")
 
         #if type(population['nkids'][0]) == float:
-        population['nkids'] = population['nkids'].astype(int)
+        population['nkids'] = population['nkids'].astype(float)
         self.population_view.update(population[['age', 'time', 'child_ages', 'nkids']])
